Fixed bug in `TermImpl::deepEquals`. (#33)
Deep equality check of terms trys to recursively check all subnodes of a
Term. But at the end of `deepEquals` it still compares the `ID` of the
Terms. I think that will never happen or is at least unintended. If the
same Term is generated two times it will never be recognized as equal
because of this.